requirements.txt
Django==1.11
Pillow==5.0.0
psycopg2==2.7.4
pytz==2018.3
